[LOWCAR] Fix lowcar type size to 1 byte (from 2)
Former-commit-id: 07355ccf27fc74130224828fe56c1184f28b8293

+++ lowcar/devices/Device/Device.cpp
@@ -37,7 +37,7 @@ void Device::loop() {
                 this->last_received_ping_time = this->curr_time;
                 // If this is the first PING received, send an ACKNOWLEDGEMENT
                 if (!this->enabled) {
-                    this->msngr->lowcar_printf("Device type %d with UID ending in %X contacted; sending ACK", this->dev_id.type, this->dev_id.uid);
+                    this->msngr->lowcar_printf("Device type %d with UID ending in %X contacted; sending ACK", (uint8_t) this->dev_id.type, this->dev_id.uid);
                     this->msngr->send_message(MessageID::ACKNOWLEDGEMENT, &(this->curr_msg), &(this->dev_id));
                     this->enabled = TRUE;
                 }
